Split AgeCounters out into its own file
