+++ lib/exports.js
@@ -33,3 +33,4 @@ module.exports.ContentRepo = (
 );
 
 module.exports.Metrics = require("./components/metrics");
+module.exports.Metrics.AgeCounters = require("./components/agecounters");
